backend/imhotep_finance/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.4.0
django-csp==4.0
python-decouple==3.8
psycopg[binary]==3.2.9
requests==2.32.4
setuptools
wheel
drf_spectacular
drf_yasg